+++ Sources/Conduit/Auth/OAuth2RequestPipelineMiddleware.swift
@@ -18,6 +18,11 @@ public struct OAuth2RequestPipelineMiddleware: RequestPipelineMiddleware {
     }
     /// Provides a grant strategy to handle a token refresh. Defaults to OAuth2TokenRefreshGrantStrategyFactory.
     public var refreshStrategyFactory: OAuth2RefreshStrategyFactory? = OAuth2TokenRefreshGrantStrategyFactory()
+    /// The maximum amount of time that multi-session locks should be issued when performing token refreshes.
+    /// Short-lived processes, such as app extensions, should typically set a lower relinquish interval to defend
+    /// against lengthy session locks when terminating mid-flight. This will allow the host process to quickly pick
+    /// up where the other process left off, if it needs to. Defaults to 30 seconds.
+    public var tokenRefreshLockRelinquishInterval: TimeInterval = 30
     let clientConfiguration: OAuth2ClientConfiguration
     let authorization: OAuth2Authorization
     let tokenStorage: OAuth2TokenStore
@@ -54,8 +59,22 @@ public struct OAuth2RequestPipelineMiddleware: RequestPipelineMiddleware {
         else if let token = token,
             token.refreshToken != nil,
             refreshStrategyFactory != nil {
+            if tokenStorage.isRefreshTokenLockedFor(client: clientConfiguration, authorization: authorization),
+                let tokenLockExpiration = tokenStorage.refreshTokenLockExpirationFor(client: clientConfiguration, authorization: authorization) {
+                logger.info("Token refresh is active in an alternate session; retrying once lock is relinquished")
+                let timeout = max(0, tokenLockExpiration.timeIntervalSinceNow)
+                OAuth2TokenRefreshCoordinator.shared.waitForRefresh(timeout: timeout) {
+                    /// If we hit the edge case where an unrelated session triggers OAuth2TokenRefreshCoordinator.endTokenRefresh(),
+                    /// then we'll recursively fall back into this branch
+                    self.prepareForTransport(request: request, completion: completion)
+                }
+                return
+            }
+            tokenStorage.lockRefreshToken(timeout: tokenRefreshLockRelinquishInterval, client: clientConfiguration, authorization: authorization)
             logger.info("Token is expired, proceeding to refresh token")
+            OAuth2TokenRefreshCoordinator.shared.beginTokenRefresh()
             refresh(token: token) { result in
+                self.tokenStorage.unlockRefreshTokenFor(client: self.clientConfiguration, authorization: self.authorization)
                 switch result {
                 case .error(let error):
                     logger.warn("There was an error refreshing the token")
@@ -69,6 +88,7 @@ public struct OAuth2RequestPipelineMiddleware: RequestPipelineMiddleware {
                     self.tokenStorage.store(token: newToken, for: self.clientConfiguration, with: self.authorization)
                     self.makeRequestByApplyingAuthorizationHeader(to: request, with: newToken, completion: completion)
                 }
+                OAuth2TokenRefreshCoordinator.shared.endTokenRefresh()
             }
         }
         else if authorization.level == .client {